chore(cli): Phase 6 — tighten entry point, split skill-hint
bin/chkit.ts: replace the module-scoped mutable _onComplete callback
with a complete() closure scoped to run(). Extract printGlobalHelp()
so the no-command branch is one call instead of an inline
try/catch/register/print. Order imports alphabetically. The catch
block at the bottom no longer needs to reach for onComplete itself —
errors thrown from run() trigger complete(1) before re-throwing.

internal-plugins/skill-hint.ts (198 lines) split into a folder:
- agent-detect.ts: AGENT_REGISTRY, AGENT_MARKERS, findAgentRoot,
  isSkillInstalledAt
- state.ts: SkillHintState, HINT_INTERVAL_MS, readSkillHintState,
  writeSkillHintState
- prompt.ts: SKILL_INSTALL_COMMAND, promptInstall, installSkill
- plugin.ts: createSkillHintPlugin assembles default deps and the
  ChxPlugin from the modules above

No behaviour change. Importers updated to the new paths.
