+++ packages/cli/src/internal-plugins/index.ts
@@ -1,6 +1,6 @@
 import type { ChxPlugin } from '../plugins.js'
 
-import { createSkillHintPlugin } from './skill-hint.js'
+import { createSkillHintPlugin } from './skill-hint/plugin.js'
 
 export function getInternalPlugins(): ChxPlugin[] {
   return [createSkillHintPlugin()]
